+++ frontend/src/pages/xray/outbounds/transport/index.ts
@@ -3,3 +3,4 @@ export { default as KcpForm } from './kcp';
 export { default as WsForm } from './ws';
 export { default as GrpcForm } from './grpc';
 export { default as HttpUpgradeForm } from './httpupgrade';
+export { default as XhttpForm } from './xhttp';
